+++ C#/TestApp/Program.cs
@@ -9,8 +9,8 @@ namespace TestApp
         {
             //var mmt = new MouseMoveTester(TestDevices.LogitechWheelMouse);
             //var mbt = new MouseButtonsTester(TestDevices.LogitechWheelMouse);
-            //var kt = new KeyboardTester(TestDevices.WyseKeyboard);
-            var kkt = new KeyboardKeyTester(TestDevices.WyseKeyboard, AhkKeys.Obj("1"), true);
+            var kt = new KeyboardTester(TestDevices.WyseKeyboard, true);
+            //var kkt = new KeyboardKeyTester(TestDevices.WyseKeyboard, AhkKeys.Obj("1"), true);
             //var tt = new TabletTester(TestDevices.ParbloIslandA609);
             //var sct = new ScanCodeTester(TestDevices.WyseKeyboard, true);
             //var sst = new SetStateTester(TestDevices.WyseKeyboard, AhkKeys.Obj("1"));